Separate PGseed into its own file
