+++ db/index.js
@@ -1,14 +1,14 @@
 const { Pool } = require("pg");
 const faker = require("faker");
 require("dotenv").config();
+const { seedPostgres } = require("./seedPostgres");
 
 const pool = new Pool({
   host: process.env.PGHOST,
   user: process.env.PGUSER,
   password: process.env.PGPASSWORD,
   database: process.env.PGHDATABASE,
-  port: process.env.PGPORT,
-  max: 30
+  port: process.env.PGPORT
 });
 
 pool.on("error", (err, client) => {
@@ -40,49 +40,6 @@ const getAllNames = async () => {
   }
 };
 
-(async function seedPostgres(loopCountInThousands, rowCount) {
-  let client = await pool.connect();
-  await client.query(`DROP TABLE IF EXISTS items;`);
-  await client.query(`CREATE TABLE items (
-    rowId SERIAL,
-    productId VARCHAR(8) NOT NULL,
-    name VARCHAR(255) NOT NULL  
-);`);
-  await client.release();
-
-  let outerLoopProgress = 0;
-
-  const createQuery = valuesArr => {
-    let query = "INSERT INTO items (productId, name) VALUES";
-    for (let i = 0; i < rowCount; i++) {
-      const fakeItem = faker.fake("{{name.firstName}} {{hacker.ingverb}}");
-      const productId = `${outerLoopProgress + (i + 1)}`
-        .toString()
-        .padStart(8, "0");
-      valuesArr.push(productId);
-      valuesArr.push(fakeItem);
-      console.log(productId);
-      query += `($${i + i + 1}, $${i + i + 2})`;
-      if (i !== rowCount - 1) {
-        query += ", ";
-      }
-    }
-    return `${query};`;
-  };
-
-  for (let i = 0; i < loopCountInThousands; i++) {
-    const values = [];
-    client = await pool.connect();
-    try {
-      const query = await createQuery(values);
-      await client.query(query, values);
-    } catch (e) {
-      console.log(e);
-    } finally {
-      client.release();
-    }
-    outerLoopProgress += 1000;
-  }
-})(100, 1000).catch(e => console.log(e.stack));
+// seedPostgres(10000, 1000).catch(e => console.log(e.stack));
 
 module.exports = { getAllNames, getOneId };
